handle toolbar button correctly
